Add width: 230px to .affix-top
If items in the sidebar nav were longer than the width of the sidebar area, the text would flow over into the content area until the page was scrolled down enough to make the menu "stick" at the top of the page.  At that point, the width would be set to 240px by the .affix class.  Adding 230px width to .affix-top to ensure long nav-menu items always wrap.